analytics: Extract analytics/views/user_activity.py.
