+++ analytics/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import include
 from django.urls import path
 
-from analytics.views.legacy import get_activity, get_user_activity
+from analytics.views.legacy import get_activity
 from analytics.views.realm_activity import get_realm_activity
 from analytics.views.stats import (
     get_chart_data,
@@ -16,6 +16,7 @@ from analytics.views.stats import (
     stats_for_remote_realm,
 )
 from analytics.views.support import support
+from analytics.views.user_activity import get_user_activity
 from zerver.lib.rest import rest_path
 
 i18n_urlpatterns = [
